Move styles out and add constants file
