+++ src/ColorPickerForm.js
@@ -3,26 +3,8 @@ import { ChromePicker } from 'react-color';
 import { TextValidator, ValidatorForm } from 'react-material-ui-form-validator';
 import { Button } from '@material-ui/core';
 import { withStyles } from '@material-ui/styles';
+import styles from './styles/ColorPickerFormStyles';
 
-const styles = {
-    picker: {
-        width: "100% !important",
-        marginTop: "2rem",
-    },
-    container: {
-        width: "100%",
-    },
-    addColor: {
-        width: "100%",
-        padding: "1rem",
-        marginTop: "1rem",
-        fontSize: "2rem",
-    },
-    colorNameInput: {
-        width: "100%",
-        height: "70px",
-    },
-};
 
 class ColorPickerForm extends Component {
     state = {
