+++ src/DraggableColorList.js
@@ -4,7 +4,7 @@ import { SortableContainer } from 'react-sortable-hoc';
 
 const DraggableColorList = SortableContainer(({ colors, removeColor }) => {
         return(
-            <div style={{ height: "100%", paddingTop: "3rem" }}>
+            <div style={{ height: "100%" }}>
                 {colors.map((color, i) => 
                   <DraggableColorBox
                     index={i}
