+++ src/NewPaletteForm.js
@@ -11,65 +11,7 @@ import DraggableColorList from './DraggableColorList';
 import { arrayMove } from 'react-sortable-hoc';
 import PaletteFormNav from './PaletteFormNav';
 import ColorPickerForm from './ColorPickerForm';
-
-const drawerWidth = 400;
-
-const styles = theme => ({
-    root: {
-      display: 'flex',
-    },
-    hide: {
-      display: 'none',
-    },
-    drawer: {
-      width: drawerWidth,
-      flexShrink: 0,
-    },
-    drawerPaper: {
-      width: drawerWidth,
-      display: 'flex',
-      alignItems: 'center',
-    },
-    drawerHeader: {
-      display: 'flex',
-      alignItems: 'center',
-      padding: '0 8px',
-      ...theme.mixins.toolbar,
-      justifyContent: 'flex-end',
-      width: "100%"
-    },
-    content: {
-      flexGrow: 1,
-      height: "calc(100vh - 64px)",
-      padding: theme.spacing.unit * 3,
-      transition: theme.transitions.create('margin', {
-        easing: theme.transitions.easing.sharp,
-        duration: theme.transitions.duration.leavingScreen,
-      }),
-      marginLeft: -drawerWidth,
-    },
-    contentShift: {
-      transition: theme.transitions.create('margin', {
-        easing: theme.transitions.easing.easeOut,
-        duration: theme.transitions.duration.enteringScreen,
-      }),
-      marginLeft: 0,
-    },
-    container: {
-      height: "100%",
-      width: "90%",
-      display: "flex",
-      flexDirection: "column",
-      justifyContent: "center",
-      alignItems: "center"
-    },
-    buttons: {
-      width: "100%",
-    },
-    button: {
-      width: "50%",
-    },
-  });  
+import styles from './styles/NewPaletteFomrStyles';
 
 class NewPaletteForm extends Component {
     static defaultProps = {
