+++ src/PaletteFormNav.js
@@ -8,48 +8,9 @@ import Typography from '@material-ui/core/Typography';
 import IconButton from '@material-ui/core/IconButton';
 import MenuIcon from '@material-ui/icons/Menu';
 import { Button } from '@material-ui/core';
-import { TextValidator, ValidatorForm } from 'react-material-ui-form-validator';
 import { Link } from 'react-router-dom';
 import PaletteMetaForm from './PaletteMetaForm';
-
-const drawerWidth = 400;
-
-const styles = theme => ({
-    root: {
-        display: "flex",
-        height: "64px",
-    },
-    appBar: {
-        transition: theme.transitions.create(['margin', 'width'], {
-          easing: theme.transitions.easing.sharp,
-          duration: theme.transitions.duration.leavingScreen,
-        }),
-        flexDirection: "row",
-        justifyContent: "space-between",
-        alignItems: "center",
-    },
-    appBarShift: {
-        width: `calc(100% - ${drawerWidth}px)`,
-        marginLeft: drawerWidth,
-        transition: theme.transitions.create(['margin', 'width'], {
-            easing: theme.transitions.easing.easeOut,
-            duration: theme.transitions.duration.enteringScreen,
-        }),
-    },
-    menuButton: {
-        marginLeft: 12,
-        marginRight: 20,
-    },
-    navBtns: {
-        marginRight: "1rem",
-        "& a": {
-            textDecoration: "none",
-        },
-    },
-    button: {
-        margin: "0 0.5rem",
-    },
-});
+import styles from './styles/PaletteFormNavStyles';
 
 class PaletteFormNav extends Component {
     state = {
