+++ src/constants.js
@@ -1,0 +1,1 @@
+export const DRAWER_WIDTH = 400;
