+++ src/styles/ColorPickerFormStyles.js
@@ -1,0 +1,21 @@
+const styles = {
+    picker: {
+        width: "100% !important",
+        marginTop: "2rem",
+    },
+    container: {
+        width: "100%",
+    },
+    addColor: {
+        width: "100%",
+        padding: "1rem",
+        marginTop: "1rem",
+        fontSize: "2rem",
+    },
+    colorNameInput: {
+        width: "100%",
+        height: "70px",
+    },
+};
+
+export default styles;
