+++ src/styles/NewPaletteFomrStyles.js
@@ -1,0 +1,62 @@
+import { DRAWER_WIDTH } from '../constants';
+
+const drawerWidth = DRAWER_WIDTH;
+
+const styles = theme => ({
+    root: {
+      display: 'flex',
+    },
+    hide: {
+      display: 'none',
+    },
+    drawer: {
+      width: drawerWidth,
+      flexShrink: 0,
+    },
+    drawerPaper: {
+      width: drawerWidth,
+      display: 'flex',
+      alignItems: 'center',
+    },
+    drawerHeader: {
+      display: 'flex',
+      alignItems: 'center',
+      padding: '0 8px',
+      ...theme.mixins.toolbar,
+      justifyContent: 'flex-end',
+      width: "100%"
+    },
+    content: {
+      flexGrow: 1,
+      height: "calc(100vh - 64px)",
+      padding: theme.spacing.unit * 3,
+      transition: theme.transitions.create('margin', {
+        easing: theme.transitions.easing.sharp,
+        duration: theme.transitions.duration.leavingScreen,
+      }),
+      marginLeft: -drawerWidth,
+    },
+    contentShift: {
+      transition: theme.transitions.create('margin', {
+        easing: theme.transitions.easing.easeOut,
+        duration: theme.transitions.duration.enteringScreen,
+      }),
+      marginLeft: 0,
+    },
+    container: {
+      height: "100%",
+      width: "90%",
+      display: "flex",
+      flexDirection: "column",
+      justifyContent: "center",
+      alignItems: "center"
+    },
+    buttons: {
+      width: "100%",
+    },
+    button: {
+      width: "50%",
+    },
+});
+
+export default styles;
