+++ src/styles/PaletteFormNavStyles.js
@@ -1,0 +1,42 @@
+import { DRAWER_WIDTH } from '../constants';
+
+const drawerWidth = DRAWER_WIDTH;
+
+const styles = theme => ({
+    root: {
+        display: "flex",
+        height: "64px",
+    },
+    appBar: {
+        transition: theme.transitions.create(['margin', 'width'], {
+          easing: theme.transitions.easing.sharp,
+          duration: theme.transitions.duration.leavingScreen,
+        }),
+        flexDirection: "row",
+        justifyContent: "space-between",
+        alignItems: "center",
+    },
+    appBarShift: {
+        width: `calc(100% - ${drawerWidth}px)`,
+        marginLeft: drawerWidth,
+        transition: theme.transitions.create(['margin', 'width'], {
+            easing: theme.transitions.easing.easeOut,
+            duration: theme.transitions.duration.enteringScreen,
+        }),
+    },
+    menuButton: {
+        marginLeft: 12,
+        marginRight: 20,
+    },
+    navBtns: {
+        marginRight: "1rem",
+        "& a": {
+            textDecoration: "none",
+        },
+    },
+    button: {
+        margin: "0 0.5rem",
+    },
+});
+
+export default styles;
